fix: fix StartNode issue and re-arrange its component
